Fix #37 - create Time and Duration classes to handle time in the simulation.

+++ visualizer/src/main/scala/com/lukewassink/visualizer/core/Message.scala
@@ -27,7 +27,7 @@ object MessageRenderer {
 
         // The portion of its journey the message has completed.
         val t =
-          (network.time.toFloat - startTime) / (endTime - startTime)
+          (network.time - startTime) / (endTime - startTime)
         MessageData(
           message,
           sender.center.interpolate(t, receiver.center)
